Initial revision: This repo is our local fork of Anorm 2.2.1.
We haven't actually changed anything about it (yet), other than to have it
reside in the package `anorm221`, rather than `anorm`. This is so that we can
upgrade Play, which uses a newer and incompatible version of Anorm without
having to figure out how to rework Possum to use either a newer version of
Anorm or rewrite it completely to use straight JDBC. (Or Apache Commons
DbUtils, as I was attempting, before I gave up as it being too much work. This
solution takes a few hours rather than numerous days.)